Layout center/scale change and other minor changes
Parameters center and scale are handled in a more systematic way.
Feedback dictated documentation changes. Minor typo fixes.
Added unit tests for placement algorithms. Cleaned up test struture.
