+++ minorminer/layout/tests/data.py
@@ -1,0 +1,374 @@
+from numpy import array
+
+precomputed_kamada_kawai = {
+    1: {
+        0: array([-1.]),
+        1: array([-0.74999997]),
+        2: array([-0.49999993]),
+        3: array([-0.2499999]),
+        4: array([0.2499999]),
+        5: array([0.49999993]),
+        6: array([0.74999997]),
+        7: array([1.])
+    },
+    2: {
+        0: array([1., -0.06320495]),
+        1: array([0.53618338, 0.48755357]),
+        2: array([-0.03438645,  0.72389135]),
+        3: array([-0.75179941,  0.66241415]),
+        4: array([-0.99999975,  0.06320475]),
+        5: array([-0.53618328, -0.48755351]),
+        6: array([0.03438642, -0.72389123]),
+        7: array([0.75179908, -0.66241412])
+    },
+    3: {
+        0: array([-0.98709905, -0.1467455, -0.14392173]),
+        1: array([0.64128769, -0.68976331, -0.36001532]),
+        2: array([0.98708649, 0.14677012, 0.14393496]),
+        3: array([-0.64126773,  0.68975957,  0.3600116]),
+        4: array([-0.12808806,  0.01386342, -0.99999759]),
+        5: array([-0.21772691, -0.85033526,  0.49607745]),
+        6: array([0.2177274,  0.85035448, -0.49608937]),
+        7: array([0.12808017, -0.01390352,  1.])
+    },
+    10: {
+        0: array([-1.,  0.14029899,  0.28687387,  0.18651374,  0.39668944,
+                  -0.43120035,  0.15087843,  0.15002928,  0.70819066, -0.70950176]),
+        1: array([0.74702881,  0.77938909,  0.09432316, -0.50056582, -0.24865196,
+                  -0.25723954,  0.60890604, -0.69002551,  0.13703349,  0.33409856]),
+        2: array([-0.22853146, -0.45075593, -0.7181411,  0.10456035,  0.60312017,
+                  0.89164998, -0.11406198, -0.36983147, -0.62121529,  0.23176585]),
+        3: array([0.48150304, -0.46892587,  0.33694644,  0.20949211, -0.75115823,
+                  -0.20321651, -0.64572577,  0.90983074, -0.2240106,  0.1436377]),
+        4: array([0.5653143, -0.13324933, -0.14883313,  0.96652688,  0.13539118,
+                  0.35343568,  0.85628352,  0.38361186,  0.45911682, -0.04684403]),
+        5: array([-0.25021981, -0.14126589, -0.79179585, -0.66333981, -0.9165724,
+                  0.12704637, -0.19481089, -0.20852614,  0.21568323, -0.62970882]),
+        6: array([-0.26440385,  0.78508537,  0.46821357,  0.64575101,  0.02804133,
+                  -0.06709947, -0.84028861, -0.50792323, -0.43359484,  0.22454071]),
+        7: array([-0.05069104, -0.51057644,  0.47241304, -0.94893846,  0.75314046,
+                  -0.41337617,  0.17881927,  0.33283447, -0.24120348,  0.45201179])
+    }
+}
+
+precomputed_chimera = {
+    2: {
+        0: array([-1.,  0.84]),
+        4: array([-0.84,  1.]),
+        5: array([-0.84,  0.92]),
+        6: array([-0.84,  0.76]),
+        7: array([-0.84,  0.68]),
+        1: array([-0.92,  0.84]),
+        2: array([-0.76,  0.84]),
+        3: array([-0.68,  0.84]),
+        32: array([-1.,  0.28]),
+        36: array([-0.84,  0.44]),
+        37: array([-0.84,  0.36]),
+        38: array([-0.84,  0.2]),
+        39: array([-0.84,  0.12]),
+        33: array([-0.92,  0.28]),
+        34: array([-0.76,  0.28]),
+        35: array([-0.68,  0.28]),
+        64: array([-1., -0.28]),
+        68: array([-0.84, -0.12]),
+        69: array([-0.84, -0.2]),
+        70: array([-0.84, -0.36]),
+        71: array([-0.84, -0.44]),
+        65: array([-0.92, -0.28]),
+        66: array([-0.76, -0.28]),
+        67: array([-0.68, -0.28]),
+        96: array([-1., -0.84]),
+        100: array([-0.84, -0.68]),
+        101: array([-0.84, -0.76]),
+        102: array([-0.84, -0.92]),
+        103: array([-0.84, -1.]),
+        97: array([-0.92, -0.84]),
+        98: array([-0.76, -0.84]),
+        99: array([-0.68, -0.84]),
+        8: array([-0.44,  0.84]),
+        12: array([-0.28,  1.]),
+        13: array([-0.28,  0.92]),
+        14: array([-0.28,  0.76]),
+        15: array([-0.28,  0.68]),
+        9: array([-0.36,  0.84]),
+        10: array([-0.2,  0.84]),
+        11: array([-0.12,  0.84]),
+        40: array([-0.44,  0.28]),
+        44: array([-0.28,  0.44]),
+        45: array([-0.28,  0.36]),
+        46: array([-0.28,  0.2]),
+        47: array([-0.28,  0.12]),
+        41: array([-0.36,  0.28]),
+        42: array([-0.2,  0.28]),
+        43: array([-0.12,  0.28]),
+        72: array([-0.44, -0.28]),
+        76: array([-0.28, -0.12]),
+        77: array([-0.28, -0.2]),
+        78: array([-0.28, -0.36]),
+        79: array([-0.28, -0.44]),
+        73: array([-0.36, -0.28]),
+        74: array([-0.2, -0.28]),
+        75: array([-0.12, -0.28]),
+        104: array([-0.44, -0.84]),
+        108: array([-0.28, -0.68]),
+        109: array([-0.28, -0.76]),
+        110: array([-0.28, -0.92]),
+        111: array([-0.28, -1.]),
+        105: array([-0.36, -0.84]),
+        106: array([-0.2, -0.84]),
+        107: array([-0.12, -0.84]),
+        16: array([0.12, 0.84]),
+        20: array([0.28, 1.]),
+        21: array([0.28, 0.92]),
+        22: array([0.28, 0.76]),
+        23: array([0.28, 0.68]),
+        17: array([0.2, 0.84]),
+        18: array([0.36, 0.84]),
+        19: array([0.44, 0.84]),
+        48: array([0.12, 0.28]),
+        52: array([0.28, 0.44]),
+        53: array([0.28, 0.36]),
+        54: array([0.28, 0.2]),
+        55: array([0.28, 0.12]),
+        49: array([0.2, 0.28]),
+        50: array([0.36, 0.28]),
+        51: array([0.44, 0.28]),
+        80: array([0.12, -0.28]),
+        84: array([0.28, -0.12]),
+        85: array([0.28, -0.2]),
+        86: array([0.28, -0.36]),
+        87: array([0.28, -0.44]),
+        81: array([0.2, -0.28]),
+        82: array([0.36, -0.28]),
+        83: array([0.44, -0.28]),
+        112: array([0.12, -0.84]),
+        116: array([0.28, -0.68]),
+        117: array([0.28, -0.76]),
+        118: array([0.28, -0.92]),
+        119: array([0.28, -1.]),
+        113: array([0.2, -0.84]),
+        114: array([0.36, -0.84]),
+        115: array([0.44, -0.84]),
+        24: array([0.68, 0.84]),
+        28: array([0.84, 1.]),
+        29: array([0.84, 0.92]),
+        30: array([0.84, 0.76]),
+        31: array([0.84, 0.68]),
+        25: array([0.76, 0.84]),
+        26: array([0.92, 0.84]),
+        27: array([1., 0.84]),
+        56: array([0.68, 0.28]),
+        60: array([0.84, 0.44]),
+        61: array([0.84, 0.36]),
+        62: array([0.84, 0.2]),
+        63: array([0.84, 0.12]),
+        57: array([0.76, 0.28]),
+        58: array([0.92, 0.28]),
+        59: array([1., 0.28]),
+        88: array([0.68, -0.28]),
+        92: array([0.84, -0.12]),
+        93: array([0.84, -0.2]),
+        94: array([0.84, -0.36]),
+        95: array([0.84, -0.44]),
+        89: array([0.76, -0.28]),
+        90: array([0.92, -0.28]),
+        91: array([1., -0.28]),
+        120: array([0.68, -0.84]),
+        124: array([0.84, -0.68]),
+        125: array([0.84, -0.76]),
+        126: array([0.84, -0.92]),
+        127: array([0.84, -1.]),
+        121: array([0.76, -0.84]),
+        122: array([0.92, -0.84]),
+        123: array([1., -0.84])
+    }
+}
+
+precomputed_random_cubic_layout = {
+    15: array([-0.29605885, -0.11441338]),
+    27: array([-0.20634156, -0.28152258]),
+    54: array([-0.32844754, -0.85720392]),
+    55: array([-0.59923838, -0.64953362]),
+    23: array([0.22306958, 0.0746885]),
+    41: array([0.41248657, 0.33828214]),
+    30: array([0.03441744, -0.84756472]),
+    51: array([0.19351707, -0.64865335]),
+    11: array([0.66808225, 0.36162903]),
+    48: array([0.33485512, 0.50666552]),
+    53: array([0.33905499, 0.64347033]),
+    59: array([0.48548501, 0.44256252]),
+    6: array([-0.80089152,  0.37560751]),
+    44: array([-1.,  0.24298214]),
+    4: array([-0.33178892,  0.52245991]),
+    5: array([-0.53250158,  0.29556071]),
+    37: array([0.18138045, 0.77722516]),
+    38: array([-0.18504518,  0.80078899]),
+    39: array([-0.51119425, -0.05158697]),
+    14: array([-0.59769715, -0.23608015]),
+    28: array([0.290619, 0.21376569]),
+    32: array([-0.03389166,  0.10164987]),
+    29: array([0.56090889, 0.55934102]),
+    7: array([-0.95676643, -0.09543276]),
+    12: array([-0.02195172,  0.54717857]),
+    50: array([-0.12096648,  0.3276822]),
+    21: array([-0.78043951, -0.37714937]),
+    13: array([-0.67814008,  0.54836334]),
+    3: array([-0.08109102, -0.37049696]),
+    56: array([-0.0878864, -0.13106446]),
+    33: array([0.78806199, 0.11036126]),
+    45: array([0.8270143, 0.28459737]),
+    24: array([0.39271777, 0.78488296]),
+    25: array([0.82422053, -0.40657747]),
+    47: array([0.6336102, -0.73974995]),
+    36: array([-0.54376614,  0.71911572]),
+    52: array([-0.86129361,  0.51767438]),
+    18: array([-0.89610764,  0.16453643]),
+    9: array([0.70906481, -0.57626658]),
+    20: array([-0.17369164, -0.89701966]),
+    26: array([0.14168721, -0.89835308]),
+    34: array([0.15070503, 0.45482215]),
+    1: array([0.14657236, -0.41409177]),
+    16: array([0.16633424, -0.20769606]),
+    19: array([0.62309296, 0.132319]),
+    57: array([-0.66709096,  0.21863119]),
+    10: array([0.84030577, -0.05363002]),
+    49: array([-0.42083627,  0.04099726]),
+    58: array([0.48891623, -0.44054773]),
+    31: array([-0.02515267,  0.30679733]),
+    2: array([-0.38608119, -0.65445218]),
+    22: array([-0.17780117,  0.61225303]),
+    17: array([0.61721472, -0.17465881]),
+    0: array([0.29030411, -0.32740284]),
+    40: array([0.46992963, -0.19551423]),
+    43: array([0.42660648, 0.05244438]),
+    42: array([-0.27242405,  0.00964043]),
+    35: array([0.35805353, -0.91769385]),
+    8: array([0.28275477, -0.15583181]),
+    46: array([-0.3264594, -0.36878777])
+}
+
+precomputed_closest = {
+    15: 76,
+    27: 74,
+    54: 105,
+    55: 99,
+    23: 55,
+    41: 51,
+    30: 112,
+    51: 116,
+    11: 56,
+    48: 52,
+    53: 23,
+    59: 51,
+    6: 37,
+    44: 32,
+    4: 44,
+    5: 40,
+    37: 17,
+    38: 10,
+    39: 72,
+    14: 67,
+    28: 54,
+    32: 43,
+    29: 60,
+    7: 68,
+    12: 44,
+    50: 43,
+    21: 70,
+    13: 36,
+    3: 75,
+    56: 75,
+    33: 63,
+    45: 57,
+    24: 18,
+    25: 95,
+    47: 120,
+    36: 8,
+    52: 36,
+    18: 38,
+    9: 124,
+    20: 106,
+    26: 112,
+    34: 52,
+    1: 87,
+    16: 81,
+    19: 56,
+    57: 35,
+    10: 92,
+    49: 47,
+    58: 83,
+    31: 43,
+    2: 108,
+    22: 15,
+    17: 88,
+    0: 86,
+    40: 83,
+    43: 55,
+    42: 47,
+    35: 114,
+    8: 84,
+    46: 78
+}
+
+precomputed_injective = {
+    15: 76,
+    27: 74,
+    54: 105,
+    55: 99,
+    23: 55,
+    41: 50,
+    30: 112,
+    51: 116,
+    11: 57,
+    48: 52,
+    53: 23,
+    59: 51,
+    6: 37,
+    44: 32,
+    4: 44,
+    5: 40,
+    37: 17,
+    38: 10,
+    39: 72,
+    14: 67,
+    28: 54,
+    32: 48,
+    29: 60,
+    7: 68,
+    12: 11,
+    50: 42,
+    21: 70,
+    13: 7,
+    3: 75,
+    56: 77,
+    33: 63,
+    45: 61,
+    24: 18,
+    25: 95,
+    47: 120,
+    36: 8,
+    52: 36,
+    18: 38,
+    9: 124,
+    20: 106,
+    26: 113,
+    34: 53,
+    1: 87,
+    16: 81,
+    19: 56,
+    57: 35,
+    10: 92,
+    49: 46,
+    58: 82,
+    31: 43,
+    2: 108,
+    22: 15,
+    17: 88,
+    0: 86,
+    40: 83,
+    43: 84,
+    42: 47,
+    35: 114,
+    8: 85,
+    46: 78
+}
